rename GPURigidRegistration to main
